# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FixMyRoad Civic Platform API",
    "description": "Enterprise OpenAPI specification for FixMyRoad AI-powered road damage reporting, spatial routing, and municipal resolution workflow platform.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Citizen Registration",
        "description": "Public registration endpoint (Forces CITIZEN role)",
        "responses": {
          "201": { "description": "Registration successful" },
          "400": { "description": "Invalid input or email already exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Authentication",
        "responses": {
          "200": { "description": "Login successful via HttpOnly cookie" },
          "401": { "description": "Invalid email or password" }
        }
      }
    },
...
  }
}